[[FEAT]] Introduce experimental `function.sent`
As a proposed extension to the language at "stage 2" of TC-39's process,
this support should be exposed via an unstable option.

This change has altered the expected errors for some previously-existing
unit tests. Because JSHint errors (unlike warnings) cannot be silenced
or ignored, this change will not effect the exit status of the JSHint
command-line interface in any case.