[[FIX]] Report "Bad assignment." in destructuring
During destructuring, the return value of the internal
`checkLeftSideAssign` function was discarded, so JSHint would fail to
report invalid assignment targets in that context.

Move the reporting of Error 031 ("Bad assignment.") into the
`checkLeftSideAssignment` to enhance consistency. Improve the
implementation of that internal function by making it aware of the
spread operator (the previous implementation considered this form
"invalid" but failed to report that).