[[FIX]] Improve support for implied closures
Some environments wrap source code in an immediately-invoked function
expression prior to evaluation. This has a number of effects on the
semantics of the code and corresponding implications for the
interpretation of linting rules.

Update the internal scope tracker to insert this "implied" closure when
linting code directed at such environments.

NOTE: This patch describes a backwards-breaking change. In prior
releases of JSHint, users were free to enable the relevant enironmental
options at any position within the source text. This approach, while
more semantically accurate, cannot be retroactively applied once parsing
has begun. The unit tests demonstrate a usage scenario that cannot be
reconciled with the legacy behavior.