[[CHORE]] Disallow ambiguity in test suite
Some tests cause the same error to be reported for different characters
on the same line. Previously, a single expectation that did not specify
a character number would satisfy all errors with the same message on the
given line.

Update the test harness to disallow the specification of identical
expected errors and to report when errors are triggered at a character
other than what is expected. Update all tests that previously relied on
this ambiguity to formally encode the expected character number and
remove the expectations that were superfluous.