[[FIX]] Downgrade parameter error to warning
It is syntactically valid for one parameter to reference a subsequent
parameter in the same list, so this should not be considered an error.
However, this is almost certainly an indication of a programming
mistake, so a warning should be emitted.

Re-use the existing W003 and corresponding `latedef` warning for this
purposes.

Note that `latedef` is disabled by default, so this change modifies a
technically-incorrect error into warning which is disabled by default:

             | latedef: false (default behavior) | latedef: true
-------------|-----------------------------------|--------------
before patch | E056                              | E056
after patch  |                                   | W003

This is unfortunate given the likelihood that the usage reflects a
mistake--it would be preferable to issue this warning by default and
require explicit opt-out. However, doing so would complicate the
behavior and meaning of the `latedef` option. In this case, the
simplicity and predictability of the API takes precedence.