[[CHORE]] Revert `yield` expression improvements
JSHint's parsing of `yield` expressions is derived from code first
contributed by GitHub user "usrbincc." That user could not be reached to
sign JSHint's Contributor License Agreement. Revert the user's
contribution to the application logic (preserving the corresponding
MIT-licensed test code) so that it may be re-written under the terms of
the Contributor License Agreement.